Aplicando cambios en el llamado a rederizado y update de las entidades - Eloy

+++ HelloWorld-Box2d/core/src/com/helloworld/box2dprueba/objetos/ItemEquipable.java
@@ -8,7 +8,7 @@ public abstract class ItemEquipable extends Entidad {
     private boolean estaEquipado;
 
     //Constructor
-    public ItemEquipable(World world, int x, int y, int width, int height, boolean isStatic, boolean fixRotation) {
+    public ItemEquipable(World world, float x, float y, int width, int height, boolean isStatic, boolean fixRotation) {
         super(world, x, y, width, height, isStatic, fixRotation);
         this.estaEquipado = false;
     }
